Migrate fjs/djs/parser to .f.mjs
Implementation and proof migrate from authored TypeScript to
JSDoc-typed JavaScript under the stage-1 migration. ParseError moves
to a sibling types.ts since fjs/djs/transpiler depends on it. Private
state-machine types (DjsStackArray, DjsStackObject, DjsStackElement,
DjsStack, ParserState, ModuleState, and the eight per-state types) stay
as _-prefixed JSDoc typedefs in the module. Updates
fjs/djs/transpiler/module.f.ts's import specifiers, and fixes stale
module.f.ts/proof.f.ts doc references across six todo docs — several
already had stale line-number citations predating this migration
(verified against the pre-migration file), so only the .f.ts -> .f.mjs
extension was corrected there; references whose line numbers still
matched (or could be confidently recomputed) were fixed exactly.
