+++ fjs/djs/todo/157.md
@@ -35,13 +35,13 @@ leaf representation at the same boundary. This keeps one tokenizer and one
 container-building state machine without discarding `NumberToken.value` before
 extended JSON or RTTI validation can inspect it.
 
-`fjs/media/json/parser/module.f.ts` and `fjs/djs/parser/module.f.ts` implement the
+`fjs/media/json/parser/module.f.ts` and `fjs/djs/parser/module.f.mjs` implement the
 *same* container-building state machine. The value-state alphabet is identical:
 
 ```ts
 // json/parser/module.f.ts:32
 status: '' | '[' | '[v' | '[,' | '{' | '{k' | '{:' | '{v' | '{,'
-// djs/parser/module.f.ts:73
+// djs/parser/module.f.mjs:66
 valueState: '' | '[' | '[v' | '[,' | '{' | '{k' | '{:' | '{v' | '{,'
 ```
