+++ fjs/djs/todo/196.md
@@ -5,7 +5,7 @@
 
 ### Problem
 
-`fjs/djs/parser/module.f.ts` defines 17 token-kind handlers
+`fjs/djs/parser/module.f.mjs` defines 17 token-kind handlers
 (`parseInitialOp`, `parseNewLineRequiredOp`, `parseExportOp`,
 `parseResultOp`, `parseConstOp`, `parseConstNameOp`, `parseImportOp`,
 `parseImportNameOp`, `parseImportFromOp`, `parseValueOp`,
@@ -50,7 +50,7 @@ non-trivia, non-eof tokens, and decorates it with the three boilerplate
 branches:
 
 ```ts
-// fjs/djs/parser/module.f.ts (private)
+// fjs/djs/parser/module.f.mjs (private)
 type CoreHandler<S> = (token: DjsToken, metadata: TokenMetadata, state: S) => ParserState
 
 const wrap =
