+++ fjs/djs/transpiler/module.f.ts
@@ -13,7 +13,8 @@ import { setReplace, at } from '../../types/ordered_map/module.f.mjs'
 import type { OrderedMap } from '../../types/ordered_map/types.ts'
 import { stringToList } from '../../text/utf16/module.f.mjs'
 import { concat as pathConcat } from '../../path/module.f.mjs'
-import { type ParseError, parseFromTokens } from '../parser/module.f.ts'
+import { parseFromTokens } from '../parser/module.f.mjs'
+import type { ParseError } from '../parser/types.ts'
 import { run } from '../ast/module.f.mjs'
 import type { AstModule } from '../ast/types.ts'
 import { foldStep, pure, step } from '../../effects/module.f.mjs'
